+++ django_alt_tests/conf/models.py
@@ -2,8 +2,8 @@ from django.db import models
 
 
 class ModelA(models.Model):
-    field_1 = models.CharField(max_length=255)
     field_2 = models.IntegerField()
+    field_1 = models.CharField(max_length=255)
 
     class Meta:
         app_label = 'conf'